requirements.txt
openai-whisper
numpy==1.24.4
moviepy>=1.0.3
tqdm
scipy